requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
torchvision==0.15.2+cu118
torchaudio
numpy==1.26.4
diffusers==0.20.1
accelerate==0.23.0
transformers==4.34.0
peft==0.5.0
ftfy==6.1.1
tensorboard==2.14.0
Jinja2==3.1.2
pillow==10.3.0